Refactor to reduce code duplication.
